Add minimal testing infrastructure and initial tests.
This project is using [shunit2] rather than another testing project such
as [bats] because this software uses a plain bourne shell on FreeBSD,
that is `/bin/sh`. Consequently, no "bashisms" should be introduced and
the test suite should always be run on a FreeBSD system which is the
target platform.

[bats]: https://github.com/bats-core/bats-core
[shunit2]: https://github.com/kward/shunit2/wiki/Coding-standards

Signed-off-by: Fletcher Nichol <fnichol@nichol.ca>